refactor(ProductPreviewScreen.kt): redundant call conversion

+++ app/src/main/java/com/hezapp/ekonomis/product_preview/presentation/ProductPreviewScreen.kt
@@ -32,9 +32,9 @@ import com.hezapp.ekonomis.core.domain.product.model.PreviewProductSummary
 import com.hezapp.ekonomis.core.presentation.component.MyBottomNavBar
 import com.hezapp.ekonomis.core.presentation.component.ResponseLoader
 import com.hezapp.ekonomis.core.presentation.routing.MyRoutes
+import com.hezapp.ekonomis.core.presentation.theme.EkonomisTheme
 import com.hezapp.ekonomis.core.presentation.utils.navigateOnce
 import com.hezapp.ekonomis.product_preview.presentation.utils.ProductPreviewUiUtils
-import com.hezapp.ekonomis.core.presentation.theme.EkonomisTheme
 import org.koin.androidx.compose.koinViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -150,7 +150,7 @@ private fun PreviewProductPreviewScreen(){
                 PreviewProductSummary(
                     id = 0,
                     ppn = 11,
-                    price = (540_000).toInt(),
+                    price = 540_000,
                     quantity = 3,
                     name = "BBQ Sauce",
                     unitType = UnitType.CARTON
